pylint fixes

+++ tests/integration/conftest.py
@@ -151,7 +151,7 @@ def fixture_kafka_server(request, session_tmppath: Path) -> Iterator[KafkaServer
         # transfer_file (primarily the server's port number)
 
         # there is an issue with pylint here, see https:/github.com/tox-dev/py-filelock/issues/102
-        with FileLock(str(lock_path_for(transfer_file))):  # pylint: disable=E0110
+        with FileLock(str(lock_path_for(transfer_file))):  # pylint: disable=abstract-class-instantiated
             if transfer_file.exists():
                 config_data = json.loads(transfer_file.read_text())
                 zk_config = ZKConfig.from_dict(config_data["zookeeper"])
@@ -290,15 +290,21 @@ def fixture_registry_async_pair(tmp_path: Path, kafka_servers: KafkaServers):
             "port": slave_port,
         },
     )
-    master_process = Popen(["python", "-m", "karapace.karapace_all", str(master_config_path)])
-    slave_process = Popen(["python", "-m", "karapace.karapace_all", str(slave_config_path)])
-    try:
-        wait_for_port(master_port)
-        wait_for_port(slave_port)
-        yield f"http://127.0.0.1:{master_port}", f"http://127.0.0.1:{slave_port}"
-    finally:
-        master_process.kill()
-        slave_process.kill()
+
+    master_process = None
+    slave_process = None
+    with ExitStack() as stack:
+        try:
+            master_process = stack.enter_context(Popen(["python", "-m", "karapace.karapace_all", str(master_config_path)]))
+            slave_process = stack.enter_context(Popen(["python", "-m", "karapace.karapace_all", str(slave_config_path)]))
+            wait_for_port(master_port)
+            wait_for_port(slave_port)
+            yield f"http://127.0.0.1:{master_port}", f"http://127.0.0.1:{slave_port}"
+        finally:
+            if master_process:
+                master_process.kill()
+            if slave_process:
+                slave_process.kill()
 
 
 @pytest.fixture(scope="function", name="registry_async")
@@ -604,7 +610,7 @@ def configure_and_start_kafka(kafka_dir: Path, zk: ZKConfig) -> Tuple[KafkaConfi
             kafka_config_path=str(config_path),
         ),
     )
-    env: Dict[bytes, bytes] = dict()
+    env: Dict[bytes, bytes] = {}
     proc = Popen(kafka_cmd, env=env)
     return config, proc
